migrate the intial part of the catalog to be `martin_core`
