+++ martin/benches/bench.rs
@@ -2,7 +2,8 @@ use async_trait::async_trait;
 use criterion::async_executor::FuturesExecutor;
 use criterion::{Criterion, criterion_group, criterion_main};
 use martin::srv::DynTileSource;
-use martin::{CatalogSourceEntry, MartinResult, Source, TileData, TileSources, UrlQuery};
+use martin::{MartinResult, Source, TileData, TileSources, UrlQuery};
+use martin_core::tiles::catalog::CatalogSourceEntry;
 use martin_tile_utils::{Encoding, Format, TileCoord, TileInfo};
 use pprof::criterion::{Output, PProfProfiler};
 use tilejson::{TileJSON, tilejson};
