+++ martin/src/lib.rs
@@ -5,9 +5,7 @@ mod config;
 pub use config::{Config, ServerState, read_config};
 
 mod source;
-pub use source::{
-    CatalogSourceEntry, Source, Tile, TileData, TileInfoSource, TileSources, UrlQuery,
-};
+pub use source::{Source, TileData, TileInfoSource, TileSources, UrlQuery};
 
 mod utils;
 pub use utils::{IdResolver, MartinError, MartinResult, NO_MAIN_CACHE};
