+++ martin/src/srv/server.rs
@@ -12,6 +12,7 @@ use futures::TryFutureExt;
 #[cfg(feature = "lambda")]
 use lambda_web::{is_running_on_lambda, run_actix_on_lambda};
 use log::error;
+use martin_core::tiles::catalog::TileCatalog;
 use serde::{Deserialize, Serialize};
 
 use crate::MartinError::BindingError;
@@ -19,7 +20,6 @@ use crate::MartinResult;
 #[cfg(feature = "webui")]
 use crate::args::WebUiMode;
 use crate::config::ServerState;
-use crate::source::TileCatalog;
 use crate::srv::config::{KEEP_ALIVE_DEFAULT, LISTEN_ADDRESSES_DEFAULT, SrvConfig};
 use crate::srv::tiles::get_tile;
 use crate::srv::tiles_info::get_source_info;
