+++ martin/src/srv/tiles.rs
@@ -9,18 +9,19 @@ use actix_web::web::{Data, Path, Query};
 use actix_web::{HttpMessage, HttpRequest, HttpResponse, Result as ActixResult, route};
 use futures::future::try_join_all;
 use log::trace;
+use martin_core::tiles::Tile;
 use martin_tile_utils::{
     Encoding, Format, TileCoord, TileInfo, decode_brotli, decode_gzip, encode_brotli, encode_gzip,
 };
 use serde::Deserialize;
 
+use crate::TileData;
 use crate::args::PreferredEncoding;
 use crate::source::{TileInfoSources, TileSources, UrlQuery};
 use crate::srv::SrvConfig;
 use crate::srv::server::map_internal_error;
 use crate::utils::cache::get_or_insert_cached_value;
 use crate::utils::{CacheKey, CacheValue, MainCache, OptMainCache};
-use crate::{Tile, TileData};
 
 static SUPPORTED_ENC: &[HeaderEnc] = &[
     HeaderEnc::gzip(),
